+++ internal/kubernetes/resources/ingress.go
@@ -3,7 +3,7 @@ package resources
 import (
 	networkingv1 "k8s.io/api/networking/v1"
 
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/workload"
+	"github.com/maxfield-allison/dnsweaver/pkg/workload"
 )
 
 // ConvertIngress converts a networking.k8s.io/v1 Ingress to a workload.Workload.