+++ internal/kubernetes/resources/ingressroute.go
@@ -6,7 +6,7 @@ import (
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/workload"
+	"github.com/maxfield-allison/dnsweaver/pkg/workload"
 )
 
 // hostMatcherRegex matches Host(`hostname`) patterns in Traefik match expressions.