+++ internal/reconciler/orphan_targetmatch_test.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"testing"
 
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/provider"
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/source"
+	"github.com/maxfield-allison/dnsweaver/pkg/provider"
+	"github.com/maxfield-allison/dnsweaver/pkg/source"
 )
 
 // noTXTCapabilities returns capabilities for a provider that doesn't support TXT records.