+++ providers/pihole/client.go
@@ -13,8 +13,8 @@ import (
 	"net/url"
 	"strings"
 
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/httputil"
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/provider"
+	"github.com/maxfield-allison/dnsweaver/pkg/httputil"
+	"github.com/maxfield-allison/dnsweaver/pkg/provider"
 )
 
 // piholeRecord represents a DNS record from Pi-hole's API.