+++ sources/nginxproxy/nginxproxy_test.go
@@ -6,8 +6,8 @@ import (
 	"log/slog"
 	"testing"
 
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/source"
-	"gitlab.bluewillows.net/root/dnsweaver/pkg/workload"
+	"github.com/maxfield-allison/dnsweaver/pkg/source"
+	"github.com/maxfield-allison/dnsweaver/pkg/workload"
 )
 
 func testLogger() *slog.Logger {